requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bcrypt==4.0.1
certifi==2025.7.14
cffi==1.17.1
click==8.3.0
colorama==0.4.6
cryptography==45.0.5
distro==1.9.0
ecdsa==0.19.1
fastapi==0.116.1
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.10
jiter==0.10.0
multidict==6.6.3
openai==1.97.1
passlib==1.7.4
propcache==0.3.2
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyodbc==5.2.0
python-dotenv==1.1.1
python-jose==3.5.0
PyYAML==6.0.3
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.37.0
watchfiles==1.1.0
websockets==15.0.1
wheel==0.45.1
yarl==1.20.1
